Update to clap 4
  - Enable the `env` feature for the clap build dependency, not just
    the normal dependency.

  - Disable more rustdoc lints.

  - Explicitly convert a `StyledStr` to a `String`.

  - `ArgEnum` and `arg_enum` are now called `ValueEnum` and
    `value_enum`, respectively.

  - Clap 4 is stricter about how arguments are designated: in argument
    groups and conflicts, they have to be designated using the field
    name, not the long option name.

  - `clap::AppSettings::DeriveDisplayOrder` has been removed and is
    now the default.

  - `parse(from_occurrences)` is deprecated in favor of `action =
    Count`, which takes a `u8`, not a `usize`.

  - `Command` no longer takes a lifetime.

+++ src/sq.rs
@@ -1,6 +1,8 @@
 #![doc(html_favicon_url = "https://docs.sequoia-pgp.org/favicon.png")]
 #![doc(html_logo_url = "https://docs.sequoia-pgp.org/logo.svg")]
 
+#![allow(rustdoc::invalid_rust_codeblocks)]
+#![allow(rustdoc::broken_intra_doc_links)]
 #![allow(rustdoc::bare_urls)]
 #![doc = include_str!(concat!(env!("OUT_DIR"), "/sq-usage.md"))]
 
@@ -770,7 +772,7 @@ fn main() -> Result<()> {
                 private_key_store,
                 input: &mut input,
                 message: output,
-                npasswords: command.symmetric,
+                npasswords: command.symmetric as usize,
                 recipients: &recipients,
                 signers: additional_secrets,
                 mode: command.mode,
